donations/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
stripe==5.4.0
PyJWT==2.6.0
requests==2.28.2
psycopg2-binary==2.9.5
gunicorn==20.1.0
